django_backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.0
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.8.6
pytz==2022.1
sqlparse==0.4.2
django-filter==24.3
pytz==2022.1
